Fix tests
https://jira.i-novus.ru/browse/EGISZNRPLATFORM-1701

+++ rdm-sync-spring-boot-starter/src/main/java/ru/i_novus/ms/rdm/sync/service/init/RdmSyncInitializer.java
@@ -63,10 +63,12 @@ public class RdmSyncInitializer {
                 .collect(Collectors.toSet());
 
         Set<Integer> mappingIdsToDelete = vmFromDb.stream()
-                .filter(versionMappingFromDb -> !versionMappingsForSync.contains(versionMappingFromDb))
+                .filter(versionMappingFromDb -> versionMappingsForSync.stream().noneMatch(versionMappingFromDb::equalsByRange))
                 .map(VersionMapping::getMappingId).collect(Collectors.toSet());
         logger.info("delete {} not actual mappings", mappingIdsToDelete.size());
-        rdmSyncDao.deleteVersionMappings(mappingIdsToDelete);
+        if (!mappingIdsToDelete.isEmpty()) {
+            rdmSyncDao.deleteVersionMappings(mappingIdsToDelete);
+        }
     }
 
     private void sourceLoaderServiceInit() {
